+++ tests/Api/BookingTest.php
@@ -263,6 +263,7 @@ class BookingTest extends AbstractBaseApiTestCase
         $res->setVideoConferenceEquipment(false);
         $res->setMonitorEquipment(false);
         $res->setCatering(false);
+        $res->setTelecoil(false);
         $res->setAcceptanceFlow(false);
         $res->setCapacity(10);
         $res->setPermissionBusinessPartner(true);