+++ CHANGELOG.md
@@ -11,6 +11,7 @@ See [keep a changelog](https://keepachangelog.com/en/1.0.0/) for information abo
 
 - Changed to import resources from endpoints instead of hooking into an external database.
 - Added day of week to formatted dates according to locale in bookings.
+- Introduced telecoil (teleslynge) to resources.
 
 ## [1.4.2] - 2025-09-15